+++ v2/logger/logger.go
@@ -15,14 +15,24 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"log/syslog"
 	"os"
 	"path"
 	"runtime"
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/tideland/goas/v1/version"
 )
+
+//--------------------
+// VERSION
+//--------------------
+
+// PackageVersion returns the version of the version package.
+func PackageVersion() version.Version {
+	return version.New(2, 1, 0)
+}
 
 //--------------------
 // LOG LEVEL
@@ -251,47 +261,6 @@ func (gl *GoLogger) Error(info, msg string) {
 // Critical logs a message at critical level.
 func (gl *GoLogger) Critical(info, msg string) {
 	log.Println("[CRITICAL]", info, msg)
-}
-
-// SysLogger
-type SysLogger struct {
-	writer *syslog.Writer
-}
-
-// NewGoLogger returns a logger implementation using the
-// Go syslog package.
-func NewSysLogger(tag string) (Logger, error) {
-	writer, err := syslog.New(syslog.LOG_DEBUG|syslog.LOG_LOCAL0, tag)
-	if err != nil {
-		log.Fatalf("cannot init syslog: %v", err)
-		return nil, err
-	}
-	return &SysLogger{writer}, nil
-}
-
-// Debug logs a message at debug level.
-func (sl *SysLogger) Debug(info, msg string) {
-	sl.writer.Debug(info + " " + msg)
-}
-
-// Info logs a message at info level.
-func (sl *SysLogger) Info(info, msg string) {
-	sl.writer.Info(info + " " + msg)
-}
-
-// Warning logs a message at warning level.
-func (sl *SysLogger) Warning(info, msg string) {
-	sl.writer.Warning(info + " " + msg)
-}
-
-// Error logs a message at error level.
-func (sl *SysLogger) Error(info, msg string) {
-	sl.writer.Err(info + " " + msg)
-}
-
-// Critical logs a message at critical level.
-func (sl *SysLogger) Critical(info, msg string) {
-	sl.writer.Crit(info + " " + msg)
 }
 
 //--------------------