+++ v2/loop/loop.go
@@ -14,7 +14,18 @@ package loop
 import (
 	"sync"
 	"time"
+
+	"github.com/tideland/goas/v1/version"
 )
+
+//--------------------
+// VERSION
+//--------------------
+
+// PackageVersion returns the version of the version package.
+func PackageVersion() version.Version {
+	return version.New(2, 1, 0)
+}
 
 //--------------------
 // LOOP